+++ authlib/little_auth/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 from django.utils.text import capfirst
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 
 class LittleAuthConfig(AppConfig):